fix(core): 64-bit the sort/select core so sort/argsort/partition/argpartition support > int.MaxValue
The byte >int.MaxValue oracle caught np.sort returning UNSORTED data (silent) and
np.partition throwing OverflowException on a 2.1-billion-element flat array. Both
traced to the value-side sort family being int-indexed end to end:

  * AxisSort.SortInPlace/ArgSortInto took the line length as int N = (int)shape[axis]
    (negative for a >2^31 axis -> sort early-returned unsorted); their radix scratch was
    managed uint[N]/ulong[N]/long[N], whose element count caps at ~2^31.
  * AxisPartition built new byte[N*elsize] (byte[] caps at ~2^31 -> the OverflowException)
    and ctx.n = (int)N; PrepKth truncated kth positions to int.
  * RadixSort and QuickSelect indexed every element with int (n, lo/hi, positions, the
    pivot stack); RadixSort's histogram was int* though one byte-bucket can hold up to N
    counts (> int range).

Converted the whole core to 64-bit:
  * RadixSort: n -> long, hist -> long* (bucket counts can exceed int), positions -> long.
    The insertion-sort fallbacks stay int (they run only for n <= 120).
  * QuickSelect: every element index (n, lo/hi/left/right/mid/p, block bounds, the pivot
    stack) -> long; block-local counters (numL/numR/startL/startR, the 0..BlockSize scan),
    npiv and depthLimit stay int. Added a thin int-taking PartitionAtMany overload that
    widens onto the stack and forwards, so the quantile IL kernels (line length always
    <= int.MaxValue) are unchanged.
  * AxisSort/AxisPartition: long N, and the radix key/idx/histogram + partition value/idx
    scratch moved to UNMANAGED NativeMemory (long-capable). The scalar comparison path
    (Half/Complex/Decimal) still uses managed Span/Array.Sort and so throws a clear
    NotSupportedException for a > int.MaxValue axis (documented, rather than truncating).
  * NDArray.unique.Radix.cs: histogram widened to long* to match RadixSort (its managed
    key buffers keep it <= int.MaxValue, unchanged otherwise).

Validated: bounded byte oracle now 49/49 on the real 2.1e9 array on net8.0 AND net10.0
(np.sort 25s, np.partition 68s, previously silent-wrong / OverflowException); a 200K
differential vs Array.Sort across the radix-scatter, block-Hoare, SortU64, NaN-last, byte
and 2-D axis paths (8/8); 638 sort/partition/quantile/median/percentile/unique/nanargmax/
lexsort unit tests; and the FuzzMatrix sort/partition/scan/nanreduce/lexsort/stat
bit-exact-vs-NumPy tiers (9/9). argsort/argpartition ride the same converted cores (their
output-shape (int)a.size truncation was fixed in 85891751).

+++ src/NumSharp.Core/Manipulation/NDArray.unique.Radix.cs
@@ -61,10 +61,10 @@ namespace NumSharp
             {
                 var kb = GC.AllocateUninitializedArray<ulong>(count);
                 var kt = GC.AllocateUninitializedArray<ulong>(count);
-                var hist = GC.AllocateUninitializedArray<int>(8 * 256);
+                var hist = GC.AllocateUninitializedArray<long>(8 * 256);
                 for (int i = 0; i < count; i++) kb[i] = ToKey64(keys[i]);
                 fixed (ulong* pk = kb, pt = kt)
-                fixed (int* ph = hist)
+                fixed (long* ph = hist)
                 {
                     ulong* r = RadixSort.SortU64(pk, pt, count, ph);
                     for (int i = 0; i < count; i++) keys[i] = FromKey64<T>(r[i]);
@@ -74,10 +74,10 @@ namespace NumSharp
             {
                 var kb = GC.AllocateUninitializedArray<uint>(count);
                 var kt = GC.AllocateUninitializedArray<uint>(count);
-                var hist = GC.AllocateUninitializedArray<int>(4 * 256);
+                var hist = GC.AllocateUninitializedArray<long>(4 * 256);
                 for (int i = 0; i < count; i++) kb[i] = ToKey32(keys[i]);
                 fixed (uint* pk = kb, pt = kt)
-                fixed (int* ph = hist)
+                fixed (long* ph = hist)
                 {
                     uint* r = RadixSort.SortU32(pk, pt, count, b, ph);
                     for (int i = 0; i < count; i++) keys[i] = FromKey32<T>(r[i]);
@@ -101,11 +101,11 @@ namespace NumSharp
                 var kb = GC.AllocateUninitializedArray<ulong>(count);
                 var kt = GC.AllocateUninitializedArray<ulong>(count);
                 var it = GC.AllocateUninitializedArray<long>(count);
-                var hist = GC.AllocateUninitializedArray<int>(8 * 256);
+                var hist = GC.AllocateUninitializedArray<long>(8 * 256);
                 for (int i = 0; i < count; i++) kb[i] = ToKey64(keys[i]);
                 fixed (ulong* pk = kb, pkt = kt)
                 fixed (long* pperm = perm, pit = it)
-                fixed (int* ph = hist)
+                fixed (long* ph = hist)
                 {
                     long* sp = RadixSort.ArgSortU64(pk, pkt, pperm, pit, count, ph, out ulong* sk);
                     for (int i = 0; i < count; i++) { keys[i] = FromKey64<T>(sk[i]); perm[i] = sp[i]; }
@@ -116,11 +116,11 @@ namespace NumSharp
                 var kb = GC.AllocateUninitializedArray<uint>(count);
                 var kt = GC.AllocateUninitializedArray<uint>(count);
                 var it = GC.AllocateUninitializedArray<long>(count);
-                var hist = GC.AllocateUninitializedArray<int>(4 * 256);
+                var hist = GC.AllocateUninitializedArray<long>(4 * 256);
                 for (int i = 0; i < count; i++) kb[i] = ToKey32(keys[i]);
                 fixed (uint* pk = kb, pkt = kt)
                 fixed (long* pperm = perm, pit = it)
-                fixed (int* ph = hist)
+                fixed (long* ph = hist)
                 {
                     long* sp = RadixSort.ArgSortU32(pk, pkt, pperm, pit, count, b, ph, out uint* sk);
                     for (int i = 0; i < count; i++) { keys[i] = FromKey32<T>(sk[i]); perm[i] = sp[i]; }
